Inputの親クラス作成。InputDate・InputPassword・InputHidden実装。
