+++ Sdx/Html/InputDate.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace Sdx.Html
+{
+  class InputDate
+  {
+  }
+}
