+++ Sdx/Sdx.csproj
@@ -81,7 +81,10 @@
     <Compile Include="Html\CheckBox.cs" />
     <Compile Include="Html\Element.cs" />
     <Compile Include="Html\CheckableGroup.cs" />
+    <Compile Include="Html\Input.cs" />
+    <Compile Include="Html\InputDate.cs" />
     <Compile Include="Html\InputHidden.cs" />
+    <Compile Include="Html\InputPassword.cs" />
     <Compile Include="Validation\Email.cs" />
     <Compile Include="Validation\Errors.cs" />
     <Compile Include="Html\Form.cs" />
